+++ lib/mnemonic_kit.dart
@@ -9,9 +9,10 @@ typedef Uint8List RandomBytes(int size);
 
 class AddressPair {
   final String address;
+  final String publicKey;
   final String privateKey;
 
-  AddressPair(this.address, this.privateKey);
+  AddressPair(this.address, this.privateKey, {this.publicKey});
 }
 
 class MnemonicKit {
@@ -21,9 +22,8 @@ class MnemonicKit {
   String genMnemonicPhrase({int strength = 128, RandomBytes randomBytes}) {
     assert(strength % 32 == 0);
 
-    Uint8List entropy = randomBytes == null
-        ? Bip39().randomBytes(strength ~/ 8)
-        : randomBytes(strength ~/ 8);
+    Uint8List entropy =
+        randomBytes == null ? Bip39().randomBytes(strength ~/ 8) : randomBytes(strength ~/ 8);
 
     return Bip39().entropyToMnemonic(HEX.encode(entropy));
   }
@@ -42,17 +42,24 @@ class MnemonicKit {
 
   /// Use mnemonic phrase to generate Base58
   ///
-  /// return [String] [base58] if mneomnic is valid
+  /// return [BIP32] if mneomnic is valid
   /// return [null] if mnemonic is invalid
-  String mnemonicToBase58(String mnemonic) {
+  BIP32 mnemonicToBIP32(String mnemonic) {
     if (!validateMnemonic(mnemonic)) return null;
 
     PBKDF2 pbkdf2 = new PBKDF2();
     Uint8List seed = pbkdf2.process(mnemonic);
     BIP32 node = BIP32.fromSeed(seed);
-    String base58 = node.toBase58();
 
-    return base58;
+    return node;
+  }
+
+  /// Use mnemonic phrase to generate Base58
+  ///
+  /// return [String] [base58] if mneomnic is valid
+  /// return [null] if mnemonic is invalid
+  String mnemonicToBase58(String mnemonic) {
+    return mnemonicToBIP32(mnemonic)?.toBase58();
   }
 
   /// Generate Address and Private Key from Base58
@@ -62,9 +69,34 @@ class MnemonicKit {
     BIP32 node = BIP32.fromBase58(base58);
     BIP32 child = node.derivePath("m/44'/60'/0'/0/$accountIdx");
 
-    final String address =
-        EthAddress().ethereumAddressFromPublicKey(child.publicKey);
+    final String address = EthAddress().ethereumAddressFromPublicKey(child.publicKey);
 
-    return AddressPair(address, "0x" + HEX.encode(child.privateKey));
+    return AddressPair(address, HEX.encode(child.privateKey),
+        publicKey: HEX.encode(child.publicKey));
+  }
+
+  /// Generate Address, Public key and Private Key from BIP32
+  ///
+  /// Return [AddressPair]
+  AddressPair genAddressPairFromBIP32(BIP32 node, {int accountIdx = 0}) {
+    BIP32 child = node.derivePath("m/44'/60'/0'/0/$accountIdx");
+
+    final String address = EthAddress().ethereumAddressFromPublicKey(child.publicKey);
+
+    return AddressPair(address, HEX.encode(child.privateKey),
+        publicKey: HEX.encode(child.publicKey));
+  }
+
+  /// Generate Address, Public key and Private Key from Mnemonic Phrase
+  ///
+  /// Return [AddressPair]
+  AddressPair mnemonicPhraseToAddressPair(
+    String mnemonic, {
+    int accountIdx = 0,
+  }) {
+    return genAddressPairFromBIP32(
+      mnemonicToBIP32(mnemonic),
+      accountIdx: accountIdx,
+    );
   }
 }