+++ test/app_authentication_kit_test.dart
@@ -10,14 +10,20 @@ void main() {
   final String invalidMPharse =
       "bargain priority menu sunny depart decide joi puppy maze course achieve deny";
   final AddressPair address1 = AddressPair(
-      '0x17D5440809303fDe49E0fb1ec2D3f93f4d3EFA12',
-      '0xf72233af8564d26c7cc5a1a3eda8bf03edcb31eca9c62c9d18097cbcd0e6e4aa');
+    '0x17D5440809303fDe49E0fb1ec2D3f93f4d3EFA12',
+    'f72233af8564d26c7cc5a1a3eda8bf03edcb31eca9c62c9d18097cbcd0e6e4aa',
+    publicKey: '02340046704c14f3ad5b18006716e8a338988b904f5fd35cc66fc91e39395bb2cb',
+  );
   final AddressPair address2 = AddressPair(
-      '0x3d7fdeEb9AAA6A7Cf34dDCbae63Fc01fA5857ba9',
-      '0x2b926ce06f57fe448d92d3f0bfd76ffdd14ee6b6aa6051f5818e024efc3114f5');
+    '0x3d7fdeEb9AAA6A7Cf34dDCbae63Fc01fA5857ba9',
+    '2b926ce06f57fe448d92d3f0bfd76ffdd14ee6b6aa6051f5818e024efc3114f5',
+    publicKey: '034d0ba6902eeaa20a5e5eae06ff99a74af983e9f42d84ace8a515226767146060',
+  );
   final AddressPair address3 = AddressPair(
-      '0x411F17f1cd87B9D9C8C9f56dd904ad8E049af3b1',
-      '0x7b86fa33a7d9fd6588f6f7574f64e0248d9e07408ea712fa59ea559a307ed301');
+    '0x411F17f1cd87B9D9C8C9f56dd904ad8E049af3b1',
+    '7b86fa33a7d9fd6588f6f7574f64e0248d9e07408ea712fa59ea559a307ed301',
+    publicKey: '03a76a9c3a19fd269f51b4a7ad77ef41d0dc81d7edb6e062331be750399ff1559e',
+  );
 
   group('Generate mnemonic phrase', () {
     test('12 words Random mnemonic phrase', () {
@@ -33,8 +39,7 @@ void main() {
     });
 
     test('Fixed 12 words mnemonic phrase by fixed random bytes', () {
-      String mPharse =
-          MnemonicKit().genMnemonicPhrase(randomBytes: testingFixedBytes);
+      String mPharse = MnemonicKit().genMnemonicPhrase(randomBytes: testingFixedBytes);
       // print("fixed words: " + mPharse);
       expect(mPharse, expectedMPharse);
     });
@@ -69,10 +74,10 @@ void main() {
     });
 
     test('Generate First Address Pair', () {
-      AddressPair addressPair =
-          MnemonicKit().genAddressPairFromBase58(expectedBase58);
+      AddressPair addressPair = MnemonicKit().genAddressPairFromBase58(expectedBase58);
       expect(addressPair.address, address1.address);
       expect(addressPair.privateKey, address1.privateKey);
+      expect(addressPair.publicKey, address1.publicKey);
     });
 
     test('Generate Second Address Pair', () {
@@ -80,19 +85,28 @@ void main() {
           MnemonicKit().genAddressPairFromBase58(expectedBase58, accountIdx: 1);
       expect(addressPair.address, address2.address);
       expect(addressPair.privateKey, address2.privateKey);
+      expect(addressPair.publicKey, address2.publicKey);
     });
     test('Generate Third Address Pair', () {
       AddressPair addressPair =
           MnemonicKit().genAddressPairFromBase58(expectedBase58, accountIdx: 2);
       expect(addressPair.address, address3.address);
       expect(addressPair.privateKey, address3.privateKey);
+      expect(addressPair.publicKey, address3.publicKey);
+    });
+
+    test('Generate Address Pair from mnemonic', () {
+      AddressPair addressPair = MnemonicKit().mnemonicPhraseToAddressPair(expectedMPharse);
+      expect(addressPair.address, address1.address);
+      expect(addressPair.privateKey, address1.privateKey);
+      expect(addressPair.publicKey, address1.publicKey);
     });
   });
 }
 
 Uint8List testingFixedBytes(int size) {
-  Uint8List fixedBytes = Uint8List.fromList(
-      [18, 181, 94, 44, 236, 179, 172, 113, 158, 13, 112, 137, 134, 40, 7, 29]);
+  Uint8List fixedBytes =
+      Uint8List.fromList([18, 181, 94, 44, 236, 179, 172, 113, 158, 13, 112, 137, 134, 40, 7, 29]);
 
   return fixedBytes;
 }